Move jit/assert.h to csrc/assertions.h
I've kept JIT_ASSERT as an alias to TORCH_ASSERT which we can use
throughout the C++ code.
