+++ torch/csrc/jit/attributes.h
@@ -171,7 +171,7 @@ private:
       return v->name == name;
     });
     if(required && it == values_.end()) {
-      ::torch::jit::barf("%s:%u: %s: required undefined attribute '%s'", __FILE__, __LINE__, __func__, symbolToString(name));
+      ::torch::barf("%s:%u: %s: required undefined attribute '%s'", __FILE__, __LINE__, __func__, symbolToString(name));
     }
     JIT_ASSERT(!required || it != values_.end());
     return it;
